+++ packages/accounts/test/config.js
@@ -1,5 +1,6 @@
-const { Worker } = require('near-workspaces');
-const fs = require('fs');
+import { Worker } from 'near-workspaces';
+import fs from 'fs';
+
 let worker;
 module.exports = async function getConfig(env) {
     switch (env) {