+++ packages/near-api-js/browser-exports.js
@@ -1,2 +1,0 @@
-window.nearApi = require('./lib/browser-index');
-window.Buffer = Buffer;