+++ packages/types/src/index.ts
@@ -1,3 +1,4 @@
 export * from './assignable';
+export * from './enum';
 export * from './errors';
 export * from './provider';