+++ packages/utils/src/logger/index.ts
@@ -1,2 +1,3 @@
+export { ConsoleLogger } from './console.logger';
 export { Logger } from './logger';
 export type { LoggerService } from './interface';