+++ packages/utils/src/provider.ts
@@ -1,7 +1,7 @@
 import { FinalExecutionOutcome } from '@near-js/types';
 
 /** @hidden */
-export function getTransactionLastResult(txResult: FinalExecutionOutcome): any {
+export function getTransactionLastResult(txResult: FinalExecutionOutcome): Exclude<object | string | number | null, Function> {
     if (typeof txResult.status === 'object' && typeof txResult.status.SuccessValue === 'string') {
         const value = Buffer.from(txResult.status.SuccessValue, 'base64').toString();
         try {