+++ packages/wallet-account/test/wallet_account.test.ts
@@ -1,0 +1,3 @@
+import { createTransactions } from './wallet_account';
+
+createTransactions();