+++ packages/keystores-node/CHANGELOG.md
@@ -1,5 +1,52 @@
 # @near-js/keystores-node
 
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies [[`5b0bbbc1`](https://github.com/near/near-api-js/commit/5b0bbbc17ffe7d89d7767e405d2ca700dc2bba40)]:
+  - @near-js/crypto@1.4.1
+  - @near-js/keystores@0.2.1
+
+## 0.1.0
+
+### Minor Changes
+
+- [#1353](https://github.com/near/near-api-js/pull/1353) [`73690557`](https://github.com/near/near-api-js/commit/73690557c8e2a74386fca62f4ae123abe0651403) Thanks [@andy-haynes](https://github.com/andy-haynes)! - Update to Node.js 20 LTS & pnpm 9.4, modularize packages, simplify dependencies, and update tests
+
+  **Breaking Changes**
+
+  - `near-api-js@5.0.0`
+
+    - The following functions are no longer exported:
+      - `logWarning`
+      - `fetchJson`
+      - the unnamed wrapped `fetch` function exported from `setup-node-fetch.ts`
+    - The browser bundle is no longer being built in version 5; for browser support please use modules
+
+  - `@near-js/providers@1.0.0`
+
+    - The following functions are no longer exported:
+      - `fetchJson`
+
+  - `@near-js/utils@1.0.0`
+    - The following functions are no longer exported:
+      - `logWarning`
+
+### Patch Changes
+
+- Updated dependencies [[`73690557`](https://github.com/near/near-api-js/commit/73690557c8e2a74386fca62f4ae123abe0651403)]:
+  - @near-js/crypto@1.4.0
+  - @near-js/keystores@0.2.0
+
+## 0.0.13
+
+### Patch Changes
+
+- Updated dependencies [[`9c7db11c`](https://github.com/near/near-api-js/commit/9c7db11c3c5c031a749dd72d5140f58056570e36), [`7d5a8244`](https://github.com/near/near-api-js/commit/7d5a8244a1683d7b5e82c4da1e40d834167a9a41)]:
+  - @near-js/keystores@0.1.0
+  - @near-js/crypto@1.3.0
+
 ## 0.0.12
 
 ### Patch Changes