+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,50 @@
 # @near-js/utils
 
+## 1.0.1
+
+### Patch Changes
+
+- [#1401](https://github.com/near/near-api-js/pull/1401) [`5b0bbbc1`](https://github.com/near/near-api-js/commit/5b0bbbc17ffe7d89d7767e405d2ca700dc2bba40) Thanks [@andy-haynes](https://github.com/andy-haynes)! - @near-js/client package
+
+- Updated dependencies [[`5b0bbbc1`](https://github.com/near/near-api-js/commit/5b0bbbc17ffe7d89d7767e405d2ca700dc2bba40)]:
+  - @near-js/types@0.3.1
+
+## 1.0.0
+
+### Major Changes
+
+- [#1353](https://github.com/near/near-api-js/pull/1353) [`73690557`](https://github.com/near/near-api-js/commit/73690557c8e2a74386fca62f4ae123abe0651403) Thanks [@andy-haynes](https://github.com/andy-haynes)! - Update to Node.js 20 LTS & pnpm 9.4, modularize packages, simplify dependencies, and update tests
+
+  **Breaking Changes**
+
+  - `near-api-js@5.0.0`
+
+    - The following functions are no longer exported:
+      - `logWarning`
+      - `fetchJson`
+      - the unnamed wrapped `fetch` function exported from `setup-node-fetch.ts`
+    - The browser bundle is no longer being built in version 5; for browser support please use modules
+
+  - `@near-js/providers@1.0.0`
+
+    - The following functions are no longer exported:
+      - `fetchJson`
+
+  - `@near-js/utils@1.0.0`
+    - The following functions are no longer exported:
+      - `logWarning`
+
+### Patch Changes
+
+- Updated dependencies [[`73690557`](https://github.com/near/near-api-js/commit/73690557c8e2a74386fca62f4ae123abe0651403)]:
+  - @near-js/types@0.3.0
+
+## 0.3.0
+
+### Minor Changes
+
+- [#1363](https://github.com/near/near-api-js/pull/1363) [`bad95007`](https://github.com/near/near-api-js/commit/bad95007edde4ed9d5989ded7f2032b9f15f5c23) Thanks [@jakmeier](https://github.com/jakmeier)! - New transaction submission errors: ShardCongested, ShardStuck, ReceiptSizeExceeded
+
 ## 0.2.2
 
 ### Patch Changes